requirements.txt
ipython
pinecone-client
numpy
pandas
plotly.express
openai
langchain
langchain_openai
langchain_experimental
langchainhub
pinecone-client
python-dotenv
tiktoken
docx2txt
pypdf
requests
aiohttp
PyYAML

